Refactor to make S-OCT installable as a git repo
