+++ .gitignore
@@ -1,0 +1,25 @@
+*
+!.venv/
+!.venv/**
+
+!CPAIOR2022/
+!CPAIOR2022/**
+
+!datasets/
+!datasets/**
+
+!s_oct/
+!s_oct/**
+
+!results/
+!results/**
+
+!scripts/
+!scripts/**
+
+!pyproject.toml
+!uv.lock
+
+!.gitignore
+
+!README.md
