+++ s_oct/decision_tree_heuristics.py
@@ -1,15 +1,17 @@
-import warnings
-from statistics import mode
-from queue import PriorityQueue
 import numpy as np
 import pandas as pd
+import warnings
+
+from queue import PriorityQueue
+from statistics import mode
 from sklearn.tree import _tree
 from sklearn.base import BaseEstimator, ClassifierMixin
 from sklearn.utils.validation import check_X_y, check_array, check_is_fitted
 from sklearn.utils.multiclass import unique_labels
 from sklearn.svm import LinearSVC, SVC
-from .l1svm import L1SVM
-from .decision_tree import UnivariateDecisionTree, MultivariateDecisionTree
+
+from s_oct.l1svm import L1SVM
+from s_oct.decision_tree import UnivariateDecisionTree, MultivariateDecisionTree
 
 def extract_sklearn_tree(sklearn_tree):
     """Convert a fitted `DecisionTreeClassifier` to a `UnivariateDecisionTree`."""
