+++ s_oct/oct_comprehensive.py
@@ -1,14 +1,16 @@
-import time
+import gurobipy as gp
 import numpy as np
 import pandas as pd
+import time
+
+from gurobipy import GRB
 from sklearn.base import BaseEstimator, ClassifierMixin
 from sklearn.utils.validation import check_X_y, check_array, check_is_fitted
 from sklearn.utils.multiclass import unique_labels
 from sklearn.tree import DecisionTreeClassifier
-import gurobipy as gp
-from gurobipy import GRB
-from .decision_tree import UnivariateDecisionTree, MultivariateDecisionTree
-from .decision_tree_heuristics import extract_sklearn_tree, MultivariateClassificationTreeHeuristic
+
+from s_oct.decision_tree import UnivariateDecisionTree, MultivariateDecisionTree
+from s_oct.decision_tree_heuristics import extract_sklearn_tree, MultivariateClassificationTreeHeuristic
 
 class OCT(ClassifierMixin, BaseEstimator):
     """Our implementation of OCT and OCT-H from Bertsimas and Dunn (2017).
