+++ scripts/datasets.py
@@ -1,9 +1,11 @@
-import warnings
 import pandas as pd
+import warnings
+
 from sklearn.base import BaseEstimator, TransformerMixin
 from sklearn.utils.validation import check_is_fitted
 from sklearn.preprocessing import KBinsDiscretizer, MinMaxScaler, OneHotEncoder
-from sklearn.compose import ColumnTransformer
+
+from scripts.consts import proj_paths
 
 class Bucketizer(TransformerMixin, BaseEstimator):
     """Bucketize numerical features.
@@ -155,7 +157,7 @@ def load_balance_scale():
     https://archive.ics.uci.edu/ml/datasets/Balance+Scale
     """
     names = ["class","left weight","left distance","right weight","right distance"]
-    df = pd.read_csv("datasets/balance-scale.data", names=names)
+    df = pd.read_csv(proj_paths.data.balance_scale, names=names)
     y = df["class"]
     X = df.drop(columns="class")
     return X, y
@@ -168,7 +170,7 @@ def load_banknote_authentication():
     https://archive.ics.uci.edu/ml/datasets/banknote+authentication
     """
     names = ["variance","skewness","curtosis","entropy","class"]
-    df = pd.read_csv("datasets/data_banknote_authentication.txt", names=names)
+    df = pd.read_csv(proj_paths.data.banknote, names=names)
     y = df["class"]
     X = df.drop(columns="class")
     return X, y
@@ -181,7 +183,7 @@ def load_blood_transfusion():
     https://archive.ics.uci.edu/ml/datasets/Blood+Transfusion+Service+Center
     """
     names = ["recency","frequency","monetary","time","class"]
-    df = pd.read_csv("datasets/transfusion.data", header=0, names=names)
+    df = pd.read_csv(proj_paths.data.transfusion, header=0, names=names)
     y = df["class"]
     X = df.drop(columns="class")
     return X, y
@@ -193,7 +195,7 @@ def load_climate_model_crashes():
     
     https://archive.ics.uci.edu/ml/datasets/climate+model+simulation+crashes
     """
-    df = pd.read_csv("datasets/pop_failures.dat", delim_whitespace=True)
+    df = pd.read_csv(proj_paths.data.pop_failures, delim_whitespace=True)
     y = df["outcome"]
     X = df.drop(columns=["Study","Run","outcome"])
     return X, y
@@ -206,7 +208,7 @@ def load_congressional_voting_records():
     https://archive.ics.uci.edu/ml/datasets/Congressional+Voting+Records
     """
     names = list("a{}".format(j+1) for j in range(17))
-    df = pd.read_csv("datasets/house-votes-84.data", names=names)
+    df = pd.read_csv(proj_paths.data.house_votes_84, names=names)
     y = df["a1"]
     X = df.drop(columns="a1")
     return X, y
@@ -218,7 +220,7 @@ def load_glass_identification():
     
     https://archive.ics.uci.edu/ml/datasets/glass+identification
     """
-    df = pd.read_csv("datasets/glass.data",
+    df = pd.read_csv(proj_paths.data.glass,
             names=["Id number","RI","Na","Mg","Al","Si","K","Ca","Ba","Fe","Type of glass"],
             index_col="Id number")
     y = df["Type of glass"]
@@ -233,12 +235,12 @@ def load_hayes_roth():
     
     https://archive.ics.uci.edu/ml/datasets/Hayes-Roth
     """
-    df = pd.read_csv("datasets/hayes-roth.data",
+    df = pd.read_csv(proj_paths.data.hayes_roth_test,
             names=["name","hobby","age","educational level","marital status","class"],
             index_col="name")
     y_train = df["class"]
     X_train = df.drop(columns="class")
-    df = pd.read_csv("datasets/hayes-roth.test",
+    df = pd.read_csv(proj_paths.data.hayes_roth_test,
             names=["hobby","age","educational level","marital status","class"])
     y_test = df["class"]
     X_test = df.drop(columns="class")
@@ -253,10 +255,10 @@ def load_image_segmentation():
     http://archive.ics.uci.edu/ml/datasets/image+segmentation
     """
     names = ["class"] + list("a{}".format(j+1) for j in range(19))
-    df = pd.read_csv("datasets/segmentation.data", skiprows=5, names=names)
+    df = pd.read_csv(proj_paths.data.segmentation_train, skiprows=5, names=names)
     y_train = df["class"]
     X_train = df.drop(columns="class")
-    df = pd.read_csv("datasets/segmentation.test", skiprows=5, names=names)
+    df = pd.read_csv(proj_paths.data.segmentation_test, skiprows=5, names=names)
     y_test = df["class"]
     X_test = df.drop(columns="class")
     return X_train, X_test, y_train, y_test
@@ -269,7 +271,7 @@ def load_ionosphere():
     https://archive.ics.uci.edu/ml/datasets/ionosphere
     """
     names = list("a{}".format(j+1) for j in range(34)) + ["class"]
-    df = pd.read_csv("datasets/ionosphere.data", names=names)
+    df = pd.read_csv(proj_paths.data.ionosphere, names=names)
     y = df["class"]
     X = df.drop(columns="class")
     return X, y
@@ -281,7 +283,7 @@ def load_parkinsons():
     
     https://archive.ics.uci.edu/ml/datasets/parkinsons
     """
-    df = pd.read_csv("datasets/parkinsons.data", index_col="name")
+    df = pd.read_csv(proj_paths.data.parkinsons, index_col="name")
     y = df["status"]
     X = df.drop(columns="status")
     return X, y
@@ -294,7 +296,7 @@ def load_soybean_small():
     https://archive.ics.uci.edu/ml/datasets/Soybean+%28Small%29
     """
     names = list("a{}".format(j+1) for j in range(35)) + ["class"]
-    df = pd.read_csv("datasets/soybean-small.data", names=names)
+    df = pd.read_csv(proj_paths.data.soybean, names=names)
     y = df["class"]
     X = df.drop(columns="class")
     return X, y
@@ -307,7 +309,7 @@ def load_tictactoe_endgame():
     https://archive.ics.uci.edu/ml/datasets/Tic-Tac-Toe+Endgame
     """
     names = list("a{}".format(j+1) for j in range(9)) + ["class"]
-    df = pd.read_csv("datasets/tic-tac-toe.data", names=names)
+    df = pd.read_csv(proj_paths.data.tic_tac_toe, names=names)
     y = df["class"]
     X = df.drop(columns="class")
     return X, y
