+++ scripts/flow_oct_comprehensive.py
@@ -1,12 +1,18 @@
-"""S-OCT comprehensive experiments."""
+"""FlowOCT comprehensive experiments."""
 from datetime import datetime
 from sklearn.datasets import load_iris, load_wine, load_breast_cancer
-from sklearn.preprocessing import MinMaxScaler, OneHotEncoder
+from sklearn.preprocessing import OneHotEncoder
 from sklearn.pipeline import Pipeline
 from sklearn.utils import shuffle
 from sklearn.model_selection import GridSearchCV, cross_validate
-from datasets import *
-from src.soct import SOCT
+
+from scripts.consts import proj_paths
+from s_oct.flow_oct import FlowOCT
+from scripts.datasets import (
+    QuantileBucketizer, 
+    load_balance_scale, load_congressional_voting_records, load_soybean_small,
+    load_banknote_authentication, load_blood_transfusion, load_ionosphere, load_parkinsons
+)
 
 categorical_datasets = [load_balance_scale, load_congressional_voting_records, load_soybean_small]
 sklearn_datasets = [load_iris, load_wine, load_breast_cancer]
@@ -15,15 +21,7 @@ datasets = categorical_datasets + numerical_datasets
 
 time_limit = 600
 
-trees = {
-    "baseline": SOCT(max_depth=2, time_limit=time_limit),
-    "benders-last-1": SOCT(max_depth=2, time_limit=time_limit, benders_nodes="last", n_benders_cuts=1),
-    "benders-last-10": SOCT(max_depth=2, time_limit=time_limit, benders_nodes="last", n_benders_cuts=10),
-    "init-1": SOCT(max_depth=2, time_limit=time_limit, n_init_cuts=1, init_cuts_max_iter=10),
-    "init-5": SOCT(max_depth=2, time_limit=time_limit, n_init_cuts=5, init_cuts_max_iter=10)
-}
-
-for model_name, tree in trees.items():
+for benders in [False, True]:
     for dataset in datasets:
         dataset_name = dataset.__name__[5:]
         for max_depth in [2, 3, 4]:
@@ -34,7 +32,7 @@ for model_name, tree in trees.items():
                 X, y = dataset(return_X_y=True, as_frame=True)
             else:
                 X, y = dataset()
-            tree.set_params(max_depth=max_depth)
+            tree = FlowOCT(max_depth=max_depth, benders=benders, time_limit=time_limit)
             if dataset in categorical_datasets:
                 pipeline = Pipeline([
                     ('pre', OneHotEncoder(drop='if_binary', sparse_output=False, handle_unknown='ignore')),
@@ -42,7 +40,7 @@ for model_name, tree in trees.items():
                 ])
             else:
                 pipeline = Pipeline([
-                    ('pre', MinMaxScaler(feature_range=(0.00001, 0.99999))),
+                    ('pre', QuantileBucketizer()),
                     ('tree', tree)
                 ])
             if max_depth == 2:
@@ -59,9 +57,10 @@ for model_name, tree in trees.items():
             estimators = cv_results['estimator']
             train_times = [e.best_estimator_.named_steps['tree'].fit_time_ for e in estimators]
             max_splits_values = [e.best_estimator_.named_steps['tree'].max_splits for e in estimators]
-            line = [f"SOCT-{model_name}", dataset_name, max_depth,
+            model_name = "FlowOCT-Benders" if benders else "FlowOCT"
+            line = [model_name, dataset_name, max_depth,
                     *train_scores, *test_scores, *train_times,
                     *max_splits_values]
             line = [str(x) for x in line]
-            with open("comprehensive.csv", 'a') as f:
+            with open(proj_paths.results.comprehensive, 'a') as f:
                 f.write(', '.join(line) + '\n')
