+++ scripts/oct_reimplementation_comprehensive.py
@@ -1,12 +1,17 @@
 """OCT (our reimplementation) comprehensive experiments."""
 from datetime import datetime
 from sklearn.datasets import load_iris, load_wine, load_breast_cancer
-from sklearn.preprocessing import OneHotEncoder
+from sklearn.preprocessing import OneHotEncoder, MinMaxScaler
 from sklearn.pipeline import Pipeline
 from sklearn.utils import shuffle
 from sklearn.model_selection import GridSearchCV, cross_validate
-from datasets import *
-from src.oct_comprehensive import OCT
+
+from scripts.consts import proj_paths
+from s_oct.oct_comprehensive import OCT
+from scripts.datasets import (
+    load_balance_scale, load_congressional_voting_records, load_soybean_small,
+    load_banknote_authentication, load_blood_transfusion, load_ionosphere, load_parkinsons
+)
 
 categorical_datasets = [load_balance_scale, load_congressional_voting_records, load_soybean_small]
 sklearn_datasets = [load_iris, load_wine, load_breast_cancer]
@@ -54,5 +59,5 @@ for dataset in datasets:
                 *train_scores, *test_scores, *train_times,
                 *max_splits_values]
         line = [str(x) for x in line]
-        with open("comprehensive.csv", 'a') as f:
+        with open(proj_paths.results.comprehensive, 'a') as f:
             f.write(', '.join(line) + '\n')
